requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.2
dill==0.3.8
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.29.2
idna==3.10
joblib==1.4.2
jsonpickle==4.0.2
kiwisolver==1.4.8
matplotlib==3.10.1
mido==1.3.3
more-itertools==10.6.0
multidict==6.1.0
multiprocess==0.70.16
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pretty_midi==0.2.10
propcache==0.3.0
pyarrow==19.0.1
pycparser==2.22
pyfluidsynth==1.3.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.2
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webcolors==24.11.1
xxhash==3.5.0
yarl==1.18.3
